requirements.txt
numpy==1.21.4
opencv-contrib-python==4.5.4.60
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0

pip~=21.3.1
setuptools~=56.0.0
pyserial~=3.5